Fix RTL injection and media query bug
Currently, StyleXSheet's injection of LTR and RTL rules isn't exclusive; the LTR rules will remain in effect alongside the RTL ones.

In addition, the prepending of the selector for targeting either LTR or RTL doesn't take into account media queries (#62).

This commit addresses both issues.

(To be clear, we're aware that stylex won't use `html[dir]` to handle RTL long-term but need it to work while it does.)