Adding a `stylex.extends` function
This new function will be the new way to spread custom presets.
This way, you will no longer need type annotations.

In order to implement this in a scalable way, I'm refactoring
the logic which is used for handling spreads entirely.

Instead of traversing and patching the AST before it is evaluated,
I'm extending the evaluation logic from Babel so that:
- It can run functions that we pass in
  - even if the functions return non-JSON-safe values
- It can handle SpreadElement within objects as long as the value being
  spread can be evaluated.

After this work, it is possible that the Babel plugin can become even simpler
where we simply evaluate CallExpressions using this custom evaluation logic
and get the correct answer.