Require importing stylex for babel plugin
Stylex currently assumes that the binding "stylex" is globally available
and is always bound to stylex.

The refactor maintained the behaviour but also allowed importing stylex
with different names.

This PR changes the behaviour so that stylex *must* be imported before it
can be used.

If Stylex is not imported, the file will not be transformed by the StyleX
Babel plugin.