Generate flat Objects from stylex.create
Style Namespaces are defined with possible nested objects for
pseudo selectors and at-rules.

So far, the output has maintained the same object structure but with the
values replaced with classNames.

This commit changes the output of `stylex.create` to be a flat object instead.

e.g. a color within a `:hover` pseudo selector will end up with a key of
`:hover--color` at the top-level of the object instead.

This change is useful for improving the performance of the `stylex(...)` function
for merging styles.