Add support for setting style values to be null (#167)
Sometimes you don't want to set a value, you just want to remove
a style that has already been set. With this change, you can now do that
by simply setting the value to be `null`.

StyleX will skip null values when generating classNames and will
leave the null untouched in the resulting style object.

When the styles are merged, the `null` value remove any className set for that
key previously but will not add a className of its own.

This will provide an alternative to the confusing values of `revert` and `unset`
and create a smaller CSS file overall.