Use stable string results (#133)
* Improved compile time resolution of `stylex()` for stable results.

The StyleX Babel Plugin will try to compile away `stylex()` calls when
all the arguments are statically known. This is a performance optimization
that allows us to avoid the runtime overhead of merging objects.

When some of the arguments are applied conditionally, the className
produced would be equivalent to the runtime result. However, the order
of the classes was not guaranteed to be the same.

This PR fixes that problem and will produce strings that are identical
to what would be produced at runtime. This is done by generating an
object of possible results and choosing the correct result based on
the conditions used within the arguments to the `stylex()` function.

This new implementation should be (slightly) faster at runtime, but
produces more code overall. This is because instead of conditionally
joining string, now every possible result is a seprate string. However,
because of the large amount of overlap among the strings, the code should
gzip well.

* Change build order

* Add option for conditional style pre-compilation