Pass through keyframes object in native shim (#324)
I am planning to develop a RN shim for keyframe animations. The shim
will rely on the keyframes object to pass through stylex.keyframes
unmodified so that I can use it to shim animationName. This PR
implements the first step which is to pass the keyframe object through
stylex.keyframes without modification.

I think we should keep the errorMsg for now since this is still
experimental.

Co-authored-by: Ken Sheedlo <ksheedlo@meta.com>