feat: add new eslint rule for flagging legacy media query/pseudo class syntax
# Summary:

This diff adds a new eslint rule for flagging usage of stylex's legacy media query/pseudo class syntax that looks like this:

```js
const styles = stylex.create({
  root: {
    width: '100%',
   '@media (min-width: 600px)': {
      width: '50%',
    }
  }
});
```

The lint messages direct developers to instead use this syntax:

```js
const styles = stylex.create({
  root: {
    width: {
      default: '100%',
      '@media (min-width: 600px)': '50%'
    }
  }
});
```

The new lint rule is notably not auto-fixable in this diff but that is something I'm going to be working on next.

# Test Plan:

I introduced new eslint "RuleTester" tests that are passing.