[babel-plugin] Move "shared" modules into "babel-plugin"
All the non-duplicate unit tests have already been moved to the babel
plugin, so the shared tests are deleted (except the "types" tests). The
source code is moved into the plugin. The few instances where
dependencies on "shared" existed are modified to avoid it.

The "cli" package was importing the hash function, instead this is
duplicated in the CLI package.

The "eslint-plugin" package was importing the rule priorities, instead
these are inlined and all pseudo-classes use the same priority for now.