feat: allow inline keyframe expressions inside viewTransitionClass
## Overview

In https://github.com/facebook/stylex/pull/1078 I added an initial implementation of `viewTransitionClass` and when I was doing some additional manual exploration I noticed that using the `keyframes` API inline in the `viewTransitionClass` object wasn't working so this diff resolves that by adding support. This is a small thing but can really help keep the view transition configuration (relatively) consise.

## Test Plan

Added unit test covering the scenario and it is passing.